LPFE-53: Moved toast to separate service
